Cleaned things up here and there

+++ src/Utils/Utils.hpp
@@ -3,8 +3,9 @@
 
 namespace Utils
 {
-    // === Example PID decoders ===
     double decodeRPM(const std::string& hexResponse);
 
-    int decodeSpeed(const std::string& hexResponse);
+    // honestly if someone goes over 256 km/h they
+    // deserve a medal
+    uint8_t decodeSpeed(const std::string& hexResponse);
 }
